# packages/api-client/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ledgerway procurement-to-payment API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type DashboardSummaryMetrics = {
  allocatedCents: number;
  committedCents: number;
  spentCents: number;
  remainingCents: number;
  pendingApprovalCount: number;
  exceptionCount: number;
};

export interface ApprovalItem {
  id: string;
  requisitionId: string;
  requisitionNumber: string;
  description: string;
  requester: string;
  amountCents: number;
  costCenter: string;
  ageHours: number;
  status: string;
}

export interface AuditEvent {
  id: string;
  actor: string;
  action: string;
  target: string;
  /** @nullable */
  metadata?: string | null;
  createdAt: string;
}

export interface DashboardSummary {
  organization: string;
  metrics: DashboardSummaryMetrics;
  attention: ApprovalItem[];
  activity: AuditEvent[];
}

export interface RequisitionInput {
  /** @minLength 3 */
  description: string;
  /** @minimum 1 */
  estimatedAmountCents: number;
  /** @minLength 1 */
  costCenter: string;
  /** @minLength 1 */
  category: string;
  neededByDate: string;
  /** @minLength 10 */
  justification: string;
  /** @nullable */
  vendorName?: string | null;
}

export interface Requisition {
  id: string;
  number: string;
  description: string;
  estimatedAmountCents: number;
  currency: string;
  costCenter: string;
  category: string;
  requester: string;
  neededByDate: string;
  status: string;
  createdAt: string;
  /** @nullable */
  currentApprover?: string | null;
  /** @nullable */
  justification?: string | null;
}

export interface ApprovalStep {
  id: string;
  order: number;
  approver: string;
  status: string;
  /** @nullable */
  decidedAt?: string | null;
  /** @nullable */
  note?: string | null;
}

export type RequisitionDetail = Requisition & {
  approvalSteps: ApprovalStep[];
};

export type ApprovalDecisionInputDecision = typeof ApprovalDecisionInputDecision[keyof typeof ApprovalDecisionInputDecision];


export const ApprovalDecisionInputDecision = {
  approve: 'approve',
  reject: 'reject',
  request_changes: 'request_changes',
} as const;

export interface ApprovalDecisionInput {
  decision: ApprovalDecisionInputDecision;
  /** @nullable */
  note?: string | null;
}

export interface PurchaseOrderLineInput {
  description: string;
  /** @minimum 0.01 */
  quantity: number;
  /** @minimum 1 */
  unitPriceCents: number;
  /** @nullable */
  glCode?: string | null;
}

export interface PurchaseOrderInput {
  /** @nullable */
  requisitionId?: string | null;
  /** @minLength 1 */
  vendorName: string;
  costCenter: string;
  /** @minItems 1 */
  lineItems: PurchaseOrderLineInput[];
}

export interface PurchaseOrder {
  id: string;
  number: string;
  vendorName: string;
  totalAmountCents: number;
  currency: string;
  status: string;
  /** @nullable */
  issuedAt: string | null;
  lineItems: PurchaseOrderLineInput[];
}

export interface VendorInput {
  /** @minLength 2 */
  name: string;
  paymentTerms: string;
  preferred?: boolean;
}

export interface Vendor {
  id: string;
  name: string;
  status: string;
  paymentTerms: string;
  spendCents: number;
  exceptionRate: number;
  onTimeRate: number;
}

export interface ReceivingInput {
  purchaseOrderId: string;
  /** @minimum 0.01 */
  quantity: number;
  conditionNotes: string;
}

export interface ReceivingRecord {
  id: string;
  purchaseOrderId: string;
  quantity: number;
  receivedAt: string;
  conditionNotes: string;
}

export interface InvoiceInput {
  vendorName: string;
  invoiceNumber: string;
  /** @nullable */
  purchaseOrderId?: string | null;
  /** @minimum 1 */
  totalAmountCents: number;
  dueDate: string;
}

export interface Invoice {
  id: string;
  invoiceNumber: string;
  vendorName: string;
  totalAmountCents: number;
  dueDate: string;
  status: string;
  matchStatus: string;
  /** @nullable */
  purchaseOrderNumber?: string | null;
}

export type MatchResultLinesItem = {
  description: string;
  invoiceQuantity: number;
  poQuantity: number;
  receivedQuantity: number;
  status: string;
};

export interface MatchResult {
  invoiceId: string;
  status: string;
  lines: MatchResultLinesItem[];
  summary: string;
}

export type ExceptionResolutionInputResolution = typeof ExceptionResolutionInputResolution[keyof typeof ExceptionResolutionInputResolution];


export const ExceptionResolutionInputResolution = {
  approve: 'approve',
  dispute: 'dispute',
  request_correction: 'request_correction',
} as const;

export interface ExceptionResolutionInput {
  resolution: ExceptionResolutionInputResolution;
  /** @minLength 5 */
  note: string;
}

export interface PaymentBatchInput {
  /** @minItems 1 */
  invoiceIds: string[];
}

export interface PaymentBatch {
  id: string;
  number: string;
  invoiceCount: number;
  totalAmountCents: number;
  status: string;
  createdAt: string;
}

export interface Budget {
  id: string;
  label: string;
  costCenter: string;
  category: string;
  allocatedCents: number;
  committedCents: number;
  spentCents: number;
  remainingCents: number;
  utilizationPct: number;
}

export type AnalyticsOverviewSpendByCategoryItem = {
  label: string;
  amountCents: number;
};

export type AnalyticsOverviewBudgetVsActualItem = {
  label: string;
  allocatedCents: number;
  spentCents: number;
};

export type AnalyticsOverviewApprovalSla = {
  averageDays: number;
  targetDays: number;
  onTimePct: number;
};

export type AnalyticsOverviewMaverickSpend = {
  amountCents: number;
  pctOfSpend: number;
};

export interface AnalyticsOverview {
  spendByCategory: AnalyticsOverviewSpendByCategoryItem[];
  budgetVsActual: AnalyticsOverviewBudgetVsActualItem[];
  approvalSla: AnalyticsOverviewApprovalSla;
  maverickSpend: AnalyticsOverviewMaverickSpend;
}

export type ApprovalPolicyRulesItem = {
  condition: string;
  chain: string[];
};

export interface ApprovalPolicy {
  id: string;
  name: string;
  version: number;
  isActive: boolean;
  rules: ApprovalPolicyRulesItem[];
}

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

export type StatusParameter = string;

export type SearchParameter = string;

export type ListRequisitionsParams = {
status?: StatusParameter;
search?: SearchParameter;
};

export type ListPurchaseOrdersParams = {
status?: StatusParameter;
search?: SearchParameter;
};

export type ListVendorsParams = {
search?: SearchParameter;
};

export type ListInvoicesParams = {
status?: StatusParameter;
search?: SearchParameter;
};