# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ledgerway procurement-to-payment API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orgs/{orgSlug}/approvals/queue:
    get:
      operationId: listApprovalQueue
      tags: [approvals]
      summary: List approval steps assigned to the current user
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      responses:
        "200":
          description: Approval queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalItem"
  /orgs/{orgSlug}/approvals/{stepId}/decide:
    post:
      operationId: decideApproval
      tags: [approvals]
      summary: Approve, reject, or request changes
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - name: stepId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionInput"
      responses:
        "200":
          description: Updated approval item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalItem"
...
  /orgs/{orgSlug}/invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice intake record
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /orgs/{orgSlug}/invoices/{invoiceId}/run-match:
    post:
      operationId: runInvoiceMatch
      tags: [invoices]
      summary: Run a three-way invoice match
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - name: invoiceId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Match result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"
  /orgs/{orgSlug}/invoices/{invoiceId}/resolve-exception:
    post:
      operationId: resolveInvoiceException
      tags: [invoices]
      summary: Resolve an invoice exception
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - name: invoiceId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExceptionResolutionInput"
      responses:
        "200":
          description: Resolved invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /orgs/{orgSlug}/payment-batches:
    get:
      operationId: listPaymentBatches
      tags: [payments]
      summary: List payment batches
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      responses:
        "200":
          description: Payment batch list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentBatch"
    post:
      operationId: createPaymentBatch
      tags: [payments]
      summary: Create an idempotent payment batch
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 8 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentBatchInput"
      responses:
        "201":
          description: Created payment batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentBatch"
  /orgs/{orgSlug}/payment-batches/{batchId}/export:
    post:
      operationId: exportPaymentBatch
      tags: [payments]
      summary: Export a payment batch to the accounting adapter
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - name: batchId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Exported batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentBatch"
...
components:
  parameters:
    OrgSlug:
      name: orgSlug
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
    Status:
      name: status
      in: query
      required: false
      schema: { type: string }
    Search:
      name: search
      in: query
      required: false
      schema: { type: string }
...
  schemas:
...
    ApprovalItem:
      type: object
      required: [id, requisitionId, requisitionNumber, description, requester, amountCents, costCenter, ageHours, status]
      properties:
        id: { type: string, format: uuid }
        requisitionId: { type: string, format: uuid }
        requisitionNumber: { type: string }
        description: { type: string }
        requester: { type: string }
        amountCents: { type: integer }
        costCenter: { type: string }
        ageHours: { type: number }
        status: { type: string }
    ApprovalDecisionInput:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approve, reject, request_changes]
        note: { type: ["string", "null"] }
...
    InvoiceInput:
      type: object
      required: [vendorName, invoiceNumber, totalAmountCents, dueDate]
      properties:
        vendorName: { type: string }
        invoiceNumber: { type: string }
        purchaseOrderId: { type: ["string", "null"], format: uuid }
        totalAmountCents: { type: integer, minimum: 1 }
        dueDate: { type: string, format: date }
    Invoice:
      type: object
      required: [id, invoiceNumber, vendorName, totalAmountCents, dueDate, status, matchStatus]
      properties:
        id: { type: string, format: uuid }
        invoiceNumber: { type: string }
        vendorName: { type: string }
        totalAmountCents: { type: integer }
        dueDate: { type: string, format: date }
        status: { type: string }
        matchStatus: { type: string }
        purchaseOrderNumber: { type: ["string", "null"] }
    MatchResult:
      type: object
      required: [invoiceId, status, lines, summary]
      properties:
        invoiceId: { type: string, format: uuid }
        status: { type: string }
        lines:
          type: array
          items:
            type: object
            required: [description, invoiceQuantity, poQuantity, receivedQuantity, status]
            properties:
              description: { type: string }
              invoiceQuantity: { type: number }
              poQuantity: { type: number }
              receivedQuantity: { type: number }
              status: { type: string }
        summary: { type: string }
    ExceptionResolutionInput:
      type: object
      required: [resolution, note]
      properties:
        resolution:
          type: string
          enum: [approve, dispute, request_correction]
        note: { type: string, minLength: 5 }
    PaymentBatchInput:
      type: object
      required: [invoiceIds]
      properties:
        invoiceIds:
          type: array
          minItems: 1
          items: { type: string, format: uuid }
    PaymentBatch:
      type: object
      required: [id, number, invoiceCount, totalAmountCents, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        number: { type: string }
        invoiceCount: { type: integer }
        totalAmountCents: { type: integer }
        status: { type: string }
        createdAt: { type: string, format: date-time }
...